requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.23.5
scipy==1.10.1
opencv-python==4.8.0.76
pillow==9.5.0
termcolor==2.3.0
pyyaml==6.0
tqdm==4.65.0
absl-py==1.4.0
imageio==2.31.1
requests==2.25.1
scikit-image==0.19.3
imageio-ffmpeg==0.4.8
piqa==1.3.1
wandb==0.15.8
pytorch-lightning==1.6.0
gin-config==0.5.0
gdown==4.7.1
ninja==1.11.1
torch_efficient_distloss==0.1.3
